display

+++ backend/controllers/user.js
@@ -53,17 +53,14 @@ exports.getProfil = (req, res) => {
         .catch(error => res.status(400).json({ error }));
 }
 
+//Suppression du profil
 exports.deleteProfil = async (req, res) => {
         const token = req.headers.authorization.split(' ')[1]; 
         result = jwt.verify(token, config.secret);
 
-        await models.Comment.destroy({ where: {
-            userId: result.id
-        }})
+        await models.Comment.destroy({ where: { userId: result.id }})
 
-        const posts = await models.Post.findAll({ where: {
-            userId: result.id
-        }})
+        const posts = await models.Post.findAll({ where: { userId: result.id }})
         posts.forEach(post => {
             const postFilename = post.imageUrl.split('/images/')[1]
             fs.unlink(`images/${postFilename}`, () => {
@@ -73,14 +70,11 @@ exports.deleteProfil = async (req, res) => {
             })
         })
 
-        const user = await models.User.findOne({ where: {
-            id: result.id
-        }})
+        const user = await models.User.findOne({ where: { id: result.id }})
         const filename = user.imageUrl.split('/images/')[1]
         fs.unlink(`images/${filename}`, () => {
             user.destroy()
             .then(() => { res.status(200).json({ message: 'Profil supprimé !'}) })
             .catch(error => res.status(500).json({ error }));
-
         }) 
     }
